+++ project/build.properties
@@ -1,1 +1,1 @@
-sbt.version = 0.13.6
+sbt.version = 0.13.9
