+++ src/main/scala/org/apache/spark/sql/streaming/ExistingDStream.scala
@@ -20,7 +20,7 @@ package org.apache.spark.sql.streaming
 import org.apache.spark.rdd.{RDD, EmptyRDD}
 import org.apache.spark.sql.Row
 import org.apache.spark.sql.catalyst.analysis.MultiInstanceRelation
-import org.apache.spark.sql.catalyst.expressions._
+import org.apache.spark.sql.catalyst.expressions.Attribute
 import org.apache.spark.sql.catalyst.plans.logical.{Statistics, LogicalPlan}
 import org.apache.spark.sql.execution.SparkPlan
 import org.apache.spark.streaming.Time
@@ -47,7 +47,15 @@ case class LogicalDStream(output: Seq[Attribute], stream: DStream[Row])
  */
 private[streaming]
 case class PhysicalDStream(output: Seq[Attribute], @transient stream: DStream[Row])
-  extends SparkPlan with StreamTrait {
+    extends SparkPlan with StreamPlan {
+  import DStreamHelper._
 
   def children = Nil
+
+  override def doExecute(): RDD[Row] = {
+    assert(validTime != null)
+    Utils.invoke(classOf[DStream[Row]], stream, "getOrCompute", (classOf[Time], validTime))
+      .asInstanceOf[Option[RDD[Row]]]
+      .getOrElse(new EmptyRDD[Row](sparkContext))
+  }
 }
