+++ src/main/scala/org/apache/spark/sql/streaming/StreamSQLParser.scala
@@ -46,7 +46,7 @@ class StreamSQLParser(streamSqlConnector: StreamSQLContext) extends SqlParser {
 
   def apply(input: String, exceptionOnError: Boolean): Option[LogicalPlan] = {
     try {
-      Some(SparkShim.sparkParse(this, input))
+      Some(parse(input))
     } catch {
       case _ if !exceptionOnError => None
       case x: Throwable => throw x
