+++ src/main/scala/org/apache/spark/sql/streaming/windowedPlans.scala
@@ -36,11 +36,22 @@ private[streaming] case class WindowedLogicalPlan(
   override def output = child.output
 }
 
+trait WindowTrait extends StreamPlan {
+  self: UnaryNode =>
+  import DStreamHelper._
+  override def execute() = {
+    assert(validTime != null)
+    Utils.invoke(classOf[DStream[Row]], stream, "getOrCompute", (classOf[Time], validTime))
+      .asInstanceOf[Option[RDD[Row]]]
+      .getOrElse(new EmptyRDD[Row](sparkContext))
+  }
+}
+
 private[streaming] case class WindowedPhysicalPlan(
     windowDuration: Duration,
     slideDuration: Option[Duration],
     child: SparkPlan)
-  extends UnaryNode with WindowTrait {
+  extends UnaryNode with StreamPlan {
 
   @transient private val wrappedStream =
     new DStream[Row](streamSqlContext.streamingContext) {
@@ -63,4 +74,12 @@ private[streaming] case class WindowedPhysicalPlan(
     .getOrElse(wrappedStream.window(windowDuration))
 
   override def output = child.output
+
+  override def doExecute(): RDD[Row] = {
+       import DStreamHelper._
+       assert(validTime != null)
+       Utils.invoke(classOf[DStream[Row]], stream, "getOrCompute", (classOf[Time], validTime))
+         .asInstanceOf[Option[RDD[Row]]]
+         .getOrElse(new EmptyRDD[Row](sparkContext))
+    }
 }
